Update ArticleResource and NewsResource for improved navigation and display
This commit modifies the ArticleResource to hide it from navigation by setting the `shouldRegisterNavigation` property to false. Additionally, the NewsResource is updated to enable text wrapping for the 'title' and 'slug' columns, enhancing the display of these fields in the table. A minor adjustment is also made to the head.blade.php file by adding a blank line for better readability.